Refactor add_kubernetes_metadata to support autodiscovery
