+++ libbeat/processors/add_kubernetes_metadata/indexers_test.go
@@ -124,38 +124,6 @@ func TestContainerIndexer(t *testing.T) {
 	assert.Equal(t, expected.String(), indexers[1].Data.String())
 }
 
-func TestFieldMatcher(t *testing.T) {
-	testCfg := map[string]interface{}{
-		"lookup_fields": []string{},
-	}
-	fieldCfg, err := common.NewConfigFrom(testCfg)
-
-	assert.Nil(t, err)
-	matcher, err := NewFieldMatcher(*fieldCfg)
-	assert.NotNil(t, err)
-
-	testCfg["lookup_fields"] = "foo"
-	fieldCfg, _ = common.NewConfigFrom(testCfg)
-
-	matcher, err = NewFieldMatcher(*fieldCfg)
-	assert.NotNil(t, matcher)
-	assert.Nil(t, err)
-
-	input := common.MapStr{
-		"foo": "bar",
-	}
-
-	out := matcher.MetadataIndex(input)
-	assert.Equal(t, out, "bar")
-
-	nonMatchInput := common.MapStr{
-		"not": "match",
-	}
-
-	out = matcher.MetadataIndex(nonMatchInput)
-	assert.Equal(t, out, "")
-}
-
 func TestFilteredGenMeta(t *testing.T) {
 	var testConfig = common.NewConfig()
 
@@ -268,50 +236,4 @@ func TestFilteredGenMetaExclusion(t *testing.T) {
 
 	ok, _ = labelMap.HasKey("x")
 	assert.Equal(t, ok, false)
-}
-
-func TestFieldFormatMatcher(t *testing.T) {
-	testCfg := map[string]interface{}{}
-	fieldCfg, err := common.NewConfigFrom(testCfg)
-
-	assert.Nil(t, err)
-	matcher, err := NewFieldFormatMatcher(*fieldCfg)
-	assert.NotNil(t, err)
-
-	testCfg["format"] = `%{[namespace]}/%{[pod]}`
-	fieldCfg, _ = common.NewConfigFrom(testCfg)
-
-	matcher, err = NewFieldFormatMatcher(*fieldCfg)
-	assert.NotNil(t, matcher)
-	assert.Nil(t, err)
-
-	event := common.MapStr{
-		"namespace": "foo",
-		"pod":       "bar",
-	}
-
-	out := matcher.MetadataIndex(event)
-	assert.Equal(t, "foo/bar", out)
-
-	event = common.MapStr{
-		"foo": "bar",
-	}
-	out = matcher.MetadataIndex(event)
-	assert.Empty(t, out)
-
-	testCfg["format"] = `%{[dimensions.namespace]}/%{[dimensions.pod]}`
-	fieldCfg, _ = common.NewConfigFrom(testCfg)
-	matcher, err = NewFieldFormatMatcher(*fieldCfg)
-	assert.NotNil(t, matcher)
-	assert.Nil(t, err)
-
-	event = common.MapStr{
-		"dimensions": common.MapStr{
-			"pod":       "bar",
-			"namespace": "foo",
-		},
-	}
-
-	out = matcher.MetadataIndex(event)
-	assert.Equal(t, "foo/bar", out)
 }
