+++ metricbeat/processor/add_kubernetes_metadata/indexers_test.go
